+++ src/Pages/Diets/index.js
@@ -1,0 +1,9 @@
+import { Diets } from "..";
+import CreateDiet from "./CreateDiet/CreateDiet";
+import ViewDiets from "./ViewDiets/ViewDiets";
+
+export {
+    Diets,
+    CreateDiet,
+    ViewDiets,
+}
